fix: HTTP response parser correctness, size limits and leaks
- gzip: complete on wire bytes (not decompressed) so multi-read gzip
  responses are no longer truncated; client_max_body_size now counts
  stored (decompressed) bytes uniformly across raw, gzip and chunked
- chunked: enforce client_max_body_size before reporting completion so
  an oversized body is rejected even when the terminator arrives in the
  same buffer
- accept HTTP/1.0 responses (previously rejected as bad request)
- parse comma-separated Transfer-Encoding values (e.g. "gzip, chunked");
  map transfer-level gzip to content-encoding so it is actually decoded
- validate Content-Length strictly: reject negative, overflow and garbage
  instead of turning atoll("-1") into a huge size_t
- free the header value placeholder before overwriting it (1-byte leak
  per response header)
- release the gzip stream in flush to avoid leaking zlib state on reset
- fix swallowed OOM in header value handling (return 0 on failure)
- guard teparser NULL dereferences and switch bytes_readed to ssize_t
- add tests/unit/test_httpresponseparser.c covering the above

+++ protocols/http/client/httpclienthandlers.c
@@ -38,7 +38,7 @@ int http_client_read(connection_t* connection) {
     httpresponseparser_set_buffer(parser, connection->buffer);
 
     while (1) {
-        int bytes_readed = connection_data_read(connection);
+        ssize_t bytes_readed = connection_data_read(connection);
 
         switch (bytes_readed) {
         case -1:
